[graphics] Take 2D rect positions as floats, not whole pixels
The rect overloads took glm::ivec2, inherited from IRenderPass::drawImage where
every caller laid out controls on integer boundaries. Two of the sites converted
from open-coded transforms do not: the minimap origin scrolls continuously under
a fixed viewport, and rounding it to whole pixels made the map jump a pixel at a
time instead of gliding.

Caught by diffing HUD regions against the same-build noise floor. The minimap is
perfectly stable run to run, so its 17.6% difference could not be animation;
everything else in the frame matched its own noise. It is pixel-identical now.

+++ include/reone/graphics/renderer/gl2d.h
@@ -56,8 +56,8 @@ public:
     void deinit() override;
 
     void drawImage(Texture &texture,
-                   const glm::ivec2 &position,
-                   const glm::ivec2 &size,
+                   const glm::vec2 &position,
+                   const glm::vec2 &size,
                    const glm::vec4 &color = glm::vec4(1.0f),
                    const glm::mat3x4 &uv = glm::mat3x4(1.0f)) override;
 
@@ -66,8 +66,8 @@ public:
                    const glm::vec4 &color = glm::vec4(1.0f),
                    const glm::mat3x4 &uv = glm::mat3x4(1.0f)) override;
 
-    void drawRect(const glm::ivec2 &position,
-                  const glm::ivec2 &size,
+    void drawRect(const glm::vec2 &position,
+                  const glm::vec2 &size,
                   const glm::vec4 &color) override;
 
     void drawFullTargetImage(Texture &texture,
